refactor code

+++ ShopmeProject/ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/user/UserRepository.java
@@ -10,6 +10,6 @@ public interface UserRepository extends CrudRepository<TaiKhoan, Integer>{
 	@Query("SELECT u FROM TaiKhoan u WHERE u.email = :email")
 	public TaiKhoan getUserByEmail(@Param("email") String email);
 	
-	public Integer countById(Integer maTK);
+	public Long countById(Integer maTK);
 	
 }
